# Conflicts:
#	Package.swift
#	Sources/Substrate/MetalBackend/MetalCaches.swift
#	Sources/Substrate/RenderGraph/ArgumentEncoder.swift
#	Sources/Substrate/RenderGraph/Enums.swift
#	Sources/Substrate/RenderGraph/Queue.swift
#	Sources/Substrate/RenderGraph/RenderGraph.swift
#	Sources/Substrate/RenderGraph/RenderPipeline.swift
#	Sources/Substrate/RenderGraph/ResourceRegistries.swift
#	Sources/SubstrateMath/Vector.swift
#	Sources/SubstrateUtilities/Allocators.swift

+++ Sources/Substrate/RenderGraph/RenderBackend.swift
@@ -406,14 +406,10 @@ public struct FunctionConstants : Hashable, Sendable {
     }
 }
 
-public enum VertexStepFunction : UInt {
+public enum VertexStepFunction : UInt8 {
     case constant
-    
     case perVertex
-    
     case perInstance
-    
     case perPatch
-    
     case perPatchControlPoint
 }
